ndb - remove compiler warnings

+++ storage/ndb/src/common/mgmcommon/ConfigRetriever.cpp
@@ -286,7 +286,7 @@ ConfigRetriever::verifyConfig(const struct ndb_mgm_configuration * conf,
     return false;
   }
 
-  if(_type != m_node_type){
+  if(_type != (unsigned int)m_node_type){
     const char *type_s, *alias_s, *type_s2, *alias_s2;
     alias_s=
       ndb_mgm_get_node_type_alias_string((enum ndb_mgm_node_type)m_node_type,
